+++ apps/example-next-basic/src/app/dynamic/page.tsx
@@ -1,7 +1,7 @@
 import { headers } from "next/headers";
 import Link from "next/link";
-import { Counter } from "@/components/Counter";
-import { List } from "@/components/List";
+import { Counter } from "@/components/counter";
+import { List } from "@/components/list";
 
 export default async function Page() {
   const headersList = await headers();
