+++ apps/example-next-basic/src/pages/pages/ssg/[id].tsx
@@ -1,7 +1,7 @@
 import type { GetServerSideProps, GetStaticPaths } from "next";
 import Link from "next/link";
-import { Counter } from "@/components/Counter";
-import { List } from "@/components/List";
+import { Counter } from "@/components/counter";
+import { List } from "@/components/list";
 
 type Props = {
   id: number;
