+++ apps/example-next-unsafe-navigation/src/components/counter.tsx
